requirements.txt
requests>=2.31
pandas>=2.0
beautifulsoup4>=4.12
lxml>=5.0
sentence-transformers>=2.2
faiss-cpu>=1.7
streamlit>=1.30
pytest>=7.0
